Reorganize most of MPI-CUDA backend to new host-transfer backend
