+++ CMakeLists.txt
@@ -10,8 +10,8 @@ cmake_minimum_required(VERSION 3.9)
 #
 
 set(ALUMINUM_VERSION_MAJOR 0)
-set(ALUMINUM_VERSION_MINOR 3)
-set(ALUMINUM_VERSION_PATCH 1)
+set(ALUMINUM_VERSION_MINOR 4)
+set(ALUMINUM_VERSION_PATCH 0)
 set(ALUMINUM_VERSION "${ALUMINUM_VERSION_MAJOR}.${ALUMINUM_VERSION_MINOR}.${ALUMINUM_VERSION_PATCH}")
 
 project(ALUMINUM VERSION ${ALUMINUM_VERSION} LANGUAGES CXX)
@@ -32,6 +32,7 @@ option(ALUMINUM_ENABLE_CUDA "Enable CUDA support." OFF)
 option(ALUMINUM_ENABLE_MPI_CUDA "Enable MPI-CUDA support." OFF)
 option(ALUMINUM_ENABLE_NCCL "Enable NCCL support." OFF)
 option(ALUMINUM_ENABLE_MPI_CUDA_RMA "Enable RMA in MPI-CUDA." OFF)
+option(ALUMINUM_ENABLE_HOST_TRANSFER "Enable host-transfer support." OFF)
 if (ALUMINUM_ENABLE_MPI_CUDA_RMA AND NOT ALUMINUM_ENABLE_MPI_CUDA)
   message(STATUS "RMA in MPI-CUDA requested; enabling MPI-CUDA support, too.")
   set(ALUMINUM_ENABLE_CUDA ON)
@@ -44,6 +45,10 @@ if (ALUMINUM_ENABLE_NCCL AND NOT ALUMINUM_ENABLE_CUDA)
   message(STATUS "NCCL support requested; enabling CUDA support, too.")
   set(ALUMINUM_ENABLE_CUDA ON)
 endif ()
+if (ALUMINUM_ENABLE_HOST_TRANSFER AND NOT ALUMINUM_ENABLE_CUDA)
+  message(STATUS "Host-transfer support requested; enabling CUDA support, too.")
+  set(ALUMINUM_ENABLE_CUDA ON)
+endif ()
 option(ALUMINUM_DEBUG_HANG_CHECK "Enable hang checking." OFF)
 option(ALUMINUM_ENABLE_NVPROF "Enable profiling via nvprof/NVTX." OFF)
 option(ALUMINUM_ENABLE_STREAM_MEM_OPS "Enable stream memory operations." OFF)
@@ -52,12 +57,14 @@ option(ALUMINUM_ENABLE_TRACE "Enable runtime tracing." OFF)
 
 if (ALUMINUM_ENABLE_CUDA
     AND NOT ALUMINUM_ENABLE_NCCL
-    AND NOT ALUMINUM_ENABLE_MPI_CUDA)
+    AND NOT ALUMINUM_ENABLE_MPI_CUDA
+    AND NOT ALUMINUM_ENABLE_HOST_TRANSFER)
   message(FATAL_ERROR
     "CUDA has been enabled without a backend. "
     "This should not happen. "
     "Please turn on \"ALUMINUM_ENABLE_NCCL\" and/or "
-    "\"ALUMINUM_ENABLE_MPI_CUDA\" and reconfigure.")
+    "\"ALUMINUM_ENABLE_MPI_CUDA\" and/or "
+    "\"ALUMINUM_ENABLE_HOST_TRANSFER\" and reconfigure.")
 endif ()
 
 if (CMAKE_BUILD_TYPE MATCHES Debug)
@@ -161,6 +168,10 @@ if (ALUMINUM_ENABLE_CUDA)
       endif ()
     endif ()
 
+    if (ALUMINUM_ENABLE_HOST_TRANSFER)
+      set(AL_HAS_HOST_TRANSFER TRUE)
+    endif ()
+
     # Add the cuda imported target
     if (NOT TARGET cuda::cuda)
       add_library(cuda::cuda INTERFACE IMPORTED)
@@ -189,15 +200,19 @@ if (ALUMINUM_ENABLE_CUDA)
     set(ALUMINUM_ENABLE_CUDA OFF)
     set(ALUMINUM_ENABLE_MPI_CUDA OFF)
     set(ALUMINUM_ENABLE_NCCL OFF)
+    set(ALUMINUM_ENABLE_HOST_TRANSFER OFF)
     set(ALUMINUM_ENABLE_NVPROF OFF)
     set(AL_HAS_CUDA FALSE)
     set(AL_HAS_MPI_CUDA FALSE)
     set(AL_HAS_NCCL FALSE)
+    set(AL_HAS_HOST_TRANSFER FALSE)
     set(AL_HAS_NVPROF OFF)
   endif ()
 
   # Check that a backend is found
-  if (NOT AL_HAS_NCCL AND NOT AL_HAS_MPI_CUDA)
+  if (NOT AL_HAS_NCCL
+      AND NOT AL_HAS_MPI_CUDA
+      AND NOT AL_HAS_HOST_TRANSFER)
     set(ALUMINUM_ENABLE_CUDA FALSE)
     set(AL_HAS_CUDA FALSE)
   endif ()
