+++ benchmark/benchmark_allgather.cpp
@@ -31,8 +31,8 @@
 #ifdef AL_HAS_NCCL
 #include "test_utils_nccl_cuda.hpp"
 #endif
-#ifdef AL_HAS_MPI_CUDA
-#include "test_utils_mpi_cuda.hpp"
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_host_transfer_cuda.hpp"
 #endif
 
 size_t start_size = 1;
@@ -115,17 +115,17 @@ int main(int argc, char** argv) {
   } else if (backend == "NCCL") {
     do_benchmark<Al::NCCLBackend>(sizes);
 #endif    
-#ifdef AL_HAS_MPI_CUDA
-  } else if (backend == "MPI-CUDA") {
-    do_benchmark<Al::MPICUDABackend>(sizes);
+#ifdef AL_HAS_HOST_TRANSFER
+  } else if (backend == "HT") {
+    do_benchmark<Al::HTBackend>(sizes);
 #endif    
   } else {
     std::cerr << "usage: " << argv[0] << " [MPI";
 #ifdef AL_HAS_NCCL
     std::cerr << " | NCCL";
 #endif
-#ifdef AL_HAS_MPI_CUDA
-    std::cerr << " | MPI-CUDA";
+#ifdef AL_HAS_HOST_TRANSFER
+    std::cerr << " | HT";
 #endif
     std::cerr << "]" << std::endl;
     return -1;
