+++ src/ht/gather.hpp
@@ -28,78 +28,92 @@
 #pragma once
 
 #include "cuda.hpp"
-#include "mpi_cuda/communicator.hpp"
+#include "ht/communicator.hpp"
 #include "progress.hpp"
 
 namespace Al {
 namespace internal {
-namespace mpi_cuda {
+namespace ht {
 
 template <typename T>
-class AllgatherAlState : public AlState {
+class GatherAlState : public AlState {
 public:
-  AllgatherAlState(const T* sendbuf, T* recvbuf, size_t count,
-                   MPICUDACommunicator& comm, cudaStream_t stream) :
+  GatherAlState(const T* sendbuf, T* recvbuf, size_t count, int root,
+                HTCommunicator& comm, cudaStream_t stream) :
     AlState(nullptr),
-    host_mem_(get_pinned_memory<T>(comm.size()*count)),
-    count_(count),
+    rank_(comm.rank()), root_(root), count_(count),
+    host_mem_(get_pinned_memory<T>(rank_ == root_ ? comm.size()*count_ : count_)),
     comm_(comm.get_comm()),
     compute_stream(comm.get_stream()) {
 
-    int const rank = comm.rank();
+    bool const i_am_root = rank_ == root_;
     bool const inplace_operation = sendbuf == recvbuf;
 
     // Transfer data from device to host and use an event to determine when it
     // completes.
-    if (inplace_operation) {
-      AL_CHECK_CUDA(cudaMemcpyAsync(host_mem_ + rank*count_,
-                                    sendbuf + rank*count_,
-                                    sizeof(T)*count_, cudaMemcpyDeviceToHost,
-                                    stream));
-    } else {
-      AL_CHECK_CUDA(cudaMemcpyAsync(host_mem_ + rank*count_,
-                                    sendbuf, sizeof(T)*count_,
+    if (i_am_root)
+      AL_CHECK_CUDA(cudaMemcpyAsync(
+                      host_mem_+rank_*count_,
+                      inplace_operation ? sendbuf+rank_*count_ : sendbuf,
+                      sizeof(T)*count, cudaMemcpyDeviceToHost, stream));
+    else
+      AL_CHECK_CUDA(cudaMemcpyAsync(host_mem_, sendbuf, sizeof(T)*count_,
                                     cudaMemcpyDeviceToHost, stream));
-    }
     d2h_event_.record(stream);
-
-    // Enqueue the kernel to wait on the host
     gpuwait_.wait(stream);
 
-    // Transfer completed buffer back to device.
-    AL_CHECK_CUDA(cudaMemcpyAsync(recvbuf, host_mem_, sizeof(T)*count*comm.size(),
-                                  cudaMemcpyHostToDevice, stream));
-    h2d_event_.record(stream);
+    if (i_am_root) {
+      // Transfer completed buffer back to device.
+      AL_CHECK_CUDA(cudaMemcpyAsync(recvbuf, host_mem_, sizeof(T)*count*comm.size(),
+                                    cudaMemcpyHostToDevice, stream));
+      h2d_event_.record(stream);
+    }
   }
 
-  ~AllgatherAlState() override {
+  ~GatherAlState() override {
     release_pinned_memory(host_mem_);
   }
 
   bool step() override {
-    if (!ag_started_) {
+    if (!gather_started_) {
       // Check if mem xfer complete
       if (d2h_event_.query()) {
-        MPI_Iallgather(MPI_IN_PLACE, count_, mpi::TypeMap<T>(),
-                       host_mem_, count_, mpi::TypeMap<T>(), comm_, &req_);
-        ag_started_ = true;
+        if (root_ == rank_) {
+          MPI_Igather(MPI_IN_PLACE, count_, mpi::TypeMap<T>(),
+                      host_mem_, count_, mpi::TypeMap<T>(),
+                      root_, comm_, &req_);
+        } else {
+          MPI_Igather(host_mem_, count_, mpi::TypeMap<T>(),
+                      host_mem_, count_, mpi::TypeMap<T>(),
+                      root_, comm_, &req_);
+          gpuwait_.signal();
+        }
+        gather_started_ = true;
       }
       else {
         return false;
       }
     }
 
-    if (!ag_done_) {
-      // Wait for the all2all to complete
+    if (!gather_done_) {
+      // Wait for the gather to complete
       int flag;
       MPI_Test(&req_, &flag, MPI_STATUS_IGNORE);
       if (flag) {
-        ag_done_ = true;
-        gpuwait_.signal();
+        gather_done_ = true;
+        if (rank_ == root_)
+          gpuwait_.signal();
+        else
+          return true;
       }
       else {
         return false;
       }
+    }
+    else if (rank_ != root_) {
+      // Paranoia, in case step() is ever called again after returning
+      // 'true' for the first time.
+      return true;
     }
 
     // Wait for host-to-device memcopy; cleanup
@@ -113,11 +127,13 @@ public:
   bool needs_completion() const override { return false; }
   void* get_compute_stream() const override { return compute_stream; }
 
-  std::string get_name() const override { return "HTAllgather"; }
+  std::string get_name() const override { return "HTGather"; }
 
 private:
-  T* host_mem_;
+  int rank_;
+  int root_;
   size_t count_;
+  T* host_mem_;
 
   cuda::GPUWait gpuwait_;
 
@@ -126,12 +142,12 @@ private:
   MPI_Comm comm_;
   MPI_Request req_ = MPI_REQUEST_NULL;
 
-  bool ag_started_ = false;
-  bool ag_done_ = false;
+  bool gather_started_ = false;
+  bool gather_done_ = false;
 
   cudaStream_t compute_stream;
 };
 
-}  // namespace mpi_cuda
+}  // namespace ht
 }  // namespace internal
 }  // namespace Al
