+++ test/test_allgather.cpp
@@ -31,8 +31,8 @@
 #ifdef AL_HAS_NCCL
 #include "test_utils_nccl_cuda.hpp"
 #endif
-#ifdef AL_HAS_MPI_CUDA
-#include "test_utils_mpi_cuda.hpp"
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_host_transfer_cuda.hpp"
 #endif
 
 #include <stdlib.h>
@@ -166,7 +166,12 @@ int main(int argc, char** argv) {
 #endif
 #ifdef AL_HAS_MPI_CUDA
   } else if (backend == "MPI-CUDA") {
-    test_correctness<Al::MPICUDABackend>();
+    std::cerr << "Allgather not supported on MPI-CUDA backend." << std::endl;
+    std::abort();
+#endif
+#ifdef AL_HAS_HOST_TRANSFER
+  } else if (backend == "HT") {
+    test_correctness<Al::HTBackend>();
 #endif
   }
