+++ test/test_multi_nballreduces.cpp
@@ -32,8 +32,8 @@
 #ifdef AL_HAS_NCCL
 #include "test_utils_nccl_cuda.hpp"
 #endif
-#ifdef AL_HAS_MPI_CUDA
-#include "test_utils_mpi_cuda.hpp"
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_host_transfer_cuda.hpp"
 #endif
 
 size_t max_size = 1<<20;
@@ -119,17 +119,17 @@ int main(int argc, char** argv) {
   } else if (backend == "NCCL") {
     test_multiple_nballreduces<Al::NCCLBackend>();
 #endif
-#ifdef AL_HAS_MPI_CUDA
-  } else if (backend == "MPI-CUDA") {
-    test_multiple_nballreduces<Al::MPICUDABackend>();
+#ifdef AL_HAS_HOST_TRANSFER
+  } else if (backend == "HT") {
+    test_multiple_nballreduces<Al::HTBackend>();
 #endif
   } else {
     std::cerr << "usage: " << argv[0] << " [MPI";
 #ifdef AL_HAS_NCCL
     std::cerr << " | NCCL";
 #endif
-#ifdef AL_HAS_MPI_CUDA
-    std::cerr << " | MPI-CUDA";
+#ifdef AL_HAS_HOST_TRANSFER
+    std::cerr << " | HT";
 #endif
     std::cerr << "]" << std::endl;
     return -1;
